copied code to seperate project
