+++ sdkext/src/main/java/login/ProjectException.java
@@ -2,7 +2,7 @@ package login;
 
 /**
  * Thrown when something fails when creating or deleting projects.
- * 
+ *
  * @author W.Pasman
  *
  */
